theoretic adding support for Touchpad (not tested on real device, buttons not assigned)

+++ bin/set-bootr.sh
@@ -3,6 +3,11 @@ mount -o remount,rw /
 cd /sbin
 busybox ln -sf init.bootr init
 cd ..
-busybox ln -sf uImage-2.6.24-palm-joplin-3430 uImage
+
+if [ -e uImage-2.6.24-palm-joplin-3430 ]; then
+    busybox ln -sf uImage-2.6.24-palm-joplin-3430 uImage
+else
+    busybox ln -sf uImage-2.6.35-palm-tenderloin uImage
+fi
 busybox sync
